focus showing level and skills macro workingg
